# Prevent warnings and errors when setting Mash keys more than once