--- a/benchmarks/mash_method_access.rb
+++ b/benchmarks/mash_method_access.rb
@@ -1,0 +1,15 @@
+$LOAD_PATH.unshift('lib')
+
+require 'hashie'
+require 'benchmark/ips'
+
+mash = Hashie::Mash.new(test: 'value')
+
+Benchmark.ips do |x|
+  x.hold!('tmp/mash_benchmark.json')
+
+  x.report('before') { mash.test }
+  x.report('after') { mash.test }
+
+  x.compare!
+end