@michaelherold 2017-02-24
@BRMatt and @jrmhaig could you please check to see if this solves the two problems you encountered?