@jrmhaig 2017-02-24
Yes, this fixes #414. Thanks.