@michaelherold 2017-02-24
Awesome, I'm glad it works for both of you. I'll get a new version out today! 🏆